+++ ingress-controller/internal/dataplane/translator/translate_httproute_test.go
@@ -217,12 +217,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 							Namespace: "default",
 							Routes: []kongstate.Route{{ // only 1 route should be created
 								Route: kong.Route{
-									Name:         kong.String("httproute.default.basic-httproute.0.0"),
-									PreserveHost: kong.Bool(true),
-									Protocols: []*string{
-										kong.String("http"),
-										kong.String("https"),
-									},
+									Name:         lo.ToPtr("httproute.default.basic-httproute.0.0"),
+									PreserveHost: lo.ToPtr(true),
+									Protocols:    nil,
 									Hosts: []*string{
 										kong.String("konghq.com"),
 										kong.String("www.konghq.com"),
@@ -294,12 +291,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 							Namespace: "default",
 							Routes: []kongstate.Route{{ // only 1 route should be created
 								Route: kong.Route{
-									Name:         kong.String("httproute.default.basic-httproute.0.0"),
-									PreserveHost: kong.Bool(true),
-									Protocols: []*string{
-										kong.String("http"),
-										kong.String("https"),
-									},
+									Name:         lo.ToPtr("httproute.default.basic-httproute.0.0"),
+									PreserveHost: lo.ToPtr(true),
+									Protocols:    nil,
 									Tags: []*string{
 										kong.String("k8s-name:basic-httproute"),
 										kong.String("k8s-namespace:default"),
@@ -373,12 +367,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 										kong.String("~/httpbin$"),
 										kong.String("/httpbin/"),
 									},
-									PreserveHost: kong.Bool(true),
-									Protocols: []*string{
-										kong.String("http"),
-										kong.String("https"),
-									},
-									StripPath: lo.ToPtr(false),
+									PreserveHost: lo.ToPtr(true),
+									Protocols:    nil,
+									StripPath:    lo.ToPtr(false),
 									Tags: []*string{
 										kong.String("k8s-name:basic-httproute"),
 										kong.String("k8s-namespace:default"),
@@ -451,12 +442,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 									Paths: []*string{
 										kong.String("~/httpbin$"),
 									},
-									PreserveHost: kong.Bool(true),
-									Protocols: []*string{
-										kong.String("http"),
-										kong.String("https"),
-									},
-									StripPath: lo.ToPtr(false),
+									PreserveHost: lo.ToPtr(true),
+									Protocols:    nil,
+									StripPath:    lo.ToPtr(false),
 									Tags: []*string{
 										kong.String("k8s-name:basic-httproute"),
 										kong.String("k8s-namespace:default"),
@@ -531,12 +519,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 									Paths: []*string{
 										kong.String("~/httpbin$"),
 									},
-									PreserveHost: kong.Bool(true),
-									Protocols: []*string{
-										kong.String("http"),
-										kong.String("https"),
-									},
-									StripPath: lo.ToPtr(false),
+									PreserveHost: lo.ToPtr(true),
+									Protocols:    nil,
+									StripPath:    lo.ToPtr(false),
 									Tags: []*string{
 										kong.String("k8s-name:basic-httproute"),
 										kong.String("k8s-namespace:default"),
@@ -621,12 +606,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/httpbin-2$"),
 											kong.String("/httpbin-2/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -715,12 +697,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 										kong.String("~/httpbin-1$"),
 										kong.String("/httpbin-1/"),
 									},
-									PreserveHost: kong.Bool(true),
-									Protocols: []*string{
-										kong.String("http"),
-										kong.String("https"),
-									},
-									StripPath: lo.ToPtr(false),
+									PreserveHost: lo.ToPtr(true),
+									Protocols:    nil,
+									StripPath:    lo.ToPtr(false),
 									Tags: []*string{
 										kong.String("k8s-name:basic-httproute"),
 										kong.String("k8s-namespace:default"),
@@ -755,12 +734,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 										kong.String("~/httpbin-2$"),
 										kong.String("/httpbin-2/"),
 									},
-									PreserveHost: kong.Bool(true),
-									Protocols: []*string{
-										kong.String("http"),
-										kong.String("https"),
-									},
-									StripPath: lo.ToPtr(false),
+									PreserveHost: lo.ToPtr(true),
+									Protocols:    nil,
+									StripPath:    lo.ToPtr(false),
 									Tags: []*string{
 										kong.String("k8s-name:basic-httproute"),
 										kong.String("k8s-namespace:default"),
@@ -875,12 +851,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/httpbin-2$"),
 											kong.String("/httpbin-2/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -918,12 +891,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/httpbin-2$"),
 											kong.String("/httpbin-2/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -1032,12 +1002,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/path-0$"),
 											kong.String("/path-0/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -1072,12 +1039,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/path-1$"),
 											kong.String("/path-1/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -1201,12 +1165,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/path-1$"),
 											kong.String("/path-1/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -1227,13 +1188,10 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/path-3$"),
 											kong.String("/path-3/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
-										Methods:   []*string{kong.String("DELETE")},
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
+										Methods:      []*string{lo.ToPtr("DELETE")},
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -1254,12 +1212,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/path-5$"),
 											kong.String("/path-5/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
 										Headers: map[string][]string{
 											"x-header-1": {"x-value-1"},
 											"x-header-2": {"x-value-2"},
@@ -1402,12 +1357,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/path-5$"),
 											kong.String("/path-5/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -1428,13 +1380,10 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/path-3$"),
 											kong.String("/path-3/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
-										Methods:   []*string{kong.String("DELETE")},
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
+										Methods:      []*string{lo.ToPtr("DELETE")},
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -1456,12 +1405,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 											kong.String("~/path-7$"),
 											kong.String("/path-7/"),
 										},
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
-										StripPath: lo.ToPtr(false),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
+										StripPath:    lo.ToPtr(false),
 										Tags: []*string{
 											kong.String("k8s-name:basic-httproute"),
 											kong.String("k8s-namespace:default"),
@@ -1555,12 +1501,9 @@ func TestIngressRulesFromHTTPRoutes(t *testing.T) {
 							Namespace: "default",
 							Routes: []kongstate.Route{{ // only 1 route should be created
 								Route: kong.Route{
-									Name:         kong.String("httproute.default.basic-httproute.0.0"),
-									PreserveHost: kong.Bool(true),
-									Protocols: []*string{
-										kong.String("http"),
-										kong.String("https"),
-									},
+									Name:         lo.ToPtr("httproute.default.basic-httproute.0.0"),
+									PreserveHost: lo.ToPtr(true),
+									Protocols:    nil,
 									Hosts: []*string{
 										kong.String("konghq.com"),
 										kong.String("www.konghq.com"),
@@ -1698,12 +1641,9 @@ func TestIngressRulesFromHTTPRoutesCombinedServicesAcrossHTTPRoutes(t *testing.T
 								// Only 1 route should be created.
 								{
 									Route: kong.Route{
-										Name:         kong.String("httproute.default.httproute-1.0.0"),
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
+										Name:         lo.ToPtr("httproute.default.httproute-1.0.0"),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
 										Paths: kong.StringSlice(
 											"~/foo$",
 											"~/bar$",
@@ -1798,12 +1738,9 @@ func TestIngressRulesFromHTTPRoutesCombinedServicesAcrossHTTPRoutes(t *testing.T
 								// 2 routes should be created for the 2 HTTPRoutes.
 								{
 									Route: kong.Route{
-										Name:         kong.String("httproute.default.httproute-1.0.0"),
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
+										Name:         lo.ToPtr("httproute.default.httproute-1.0.0"),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
 										Paths: kong.StringSlice(
 											"~/foo$",
 										),
@@ -1820,12 +1757,9 @@ func TestIngressRulesFromHTTPRoutesCombinedServicesAcrossHTTPRoutes(t *testing.T
 								},
 								{
 									Route: kong.Route{
-										Name:         kong.String("httproute.default.httproute-2.0.0"),
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
+										Name:         lo.ToPtr("httproute.default.httproute-2.0.0"),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
 										Paths: kong.StringSlice(
 											"~/bar$",
 										),
@@ -1942,12 +1876,9 @@ func TestIngressRulesFromHTTPRoutesCombinedServicesAcrossHTTPRoutes(t *testing.T
 							Routes: []kongstate.Route{
 								{
 									Route: kong.Route{
-										Name:         kong.String("httproute.default.httproute-1.0.0"),
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
+										Name:         lo.ToPtr("httproute.default.httproute-1.0.0"),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
 										Paths: kong.StringSlice(
 											"~/foo$",
 										),
@@ -1982,12 +1913,9 @@ func TestIngressRulesFromHTTPRoutesCombinedServicesAcrossHTTPRoutes(t *testing.T
 							Routes: []kongstate.Route{
 								{
 									Route: kong.Route{
-										Name:         kong.String("httproute.ns-alter.httproute-2.0.0"),
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
+										Name:         lo.ToPtr("httproute.ns-alter.httproute-2.0.0"),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
 										Paths: kong.StringSlice(
 											"~/bar$",
 										),
@@ -2064,12 +1992,9 @@ func TestIngressRulesFromHTTPRoutesCombinedServicesAcrossHTTPRoutes(t *testing.T
 								// 2 routes where one for each rule should be created because they have different filters.
 								{
 									Route: kong.Route{
-										Name:         kong.String("httproute.default.httproute-1.0.0"),
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
+										Name:         lo.ToPtr("httproute.default.httproute-1.0.0"),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
 										Paths: kong.StringSlice(
 											"~/kong$",
 										),
@@ -2118,12 +2043,9 @@ func TestIngressRulesFromHTTPRoutesCombinedServicesAcrossHTTPRoutes(t *testing.T
 								},
 								{
 									Route: kong.Route{
-										Name:         kong.String("httproute.default.httproute-1.1.0"),
-										PreserveHost: kong.Bool(true),
-										Protocols: []*string{
-											kong.String("http"),
-											kong.String("https"),
-										},
+										Name:         lo.ToPtr("httproute.default.httproute-1.1.0"),
+										PreserveHost: lo.ToPtr(true),
+										Protocols:    nil,
 										Paths: kong.StringSlice(
 											"~/kuma$",
 										),
@@ -2259,12 +2181,9 @@ func TestIngressRulesFromHTTPRoutes_RegexPrefix(t *testing.T) {
 									Paths: []*string{
 										kong.String("~/httpbin$"),
 									},
-									PreserveHost: kong.Bool(true),
-									Protocols: []*string{
-										kong.String("http"),
-										kong.String("https"),
-									},
-									StripPath: lo.ToPtr(false),
+									PreserveHost: lo.ToPtr(true),
+									Protocols:    nil,
+									StripPath:    lo.ToPtr(false),
 									Tags: []*string{
 										kong.String("k8s-name:basic-httproute"),
 										kong.String("k8s-namespace:default"),